@@ -2,7 +2,7 @@ name: github_search
 description: A new flutter project.
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
   flutter:
