@@ -31,7 +31,8 @@ class FlutterReduxApp extends StatelessWidget {
   final Store<int> store;
   final String title;
 
-  FlutterReduxApp({Key key, this.store, this.title}) : super(key: key);
+  FlutterReduxApp({Key? key, required this.store, required this.title})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
